refactor(#133): userStore 파일 분리
