+++ src/app/_components/Header/UserProfileLinks.tsx
@@ -1,10 +1,10 @@
 import Link from 'next/link';
 import { useState } from 'react';
+import AuthModal from '@/app/_components/Header/Auth/AuthModal';
 import { dummyUserInfo } from '@/constants/dummy';
 import { AlarmSVG, CommentSVG, SearchSVG } from '@/icons/svg';
 import useSession from '@/lib/useSession';
 import Profile from './Profile';
-import AuthModal from '@/app/_components/Header/Auth/AuthModal';
 
 const UserProfileLinks = () => {
   const { alarmCount, commentCount } = dummyUserInfo;
